fix linting issues
