+++ src/Base.php
@@ -180,7 +180,7 @@ abstract class Base
     /**
      * TCPDF version.
      */
-    protected string $version = '8.6.0';
+    protected string $version = '8.6.2';
 
     /**
      * Time is seconds since EPOCH when the document was created.
