+++ src/MetaInfo.php
@@ -316,7 +316,8 @@ abstract class MetaInfo extends \Com\Tecnick\Pdf\HTML
 
         // @codingStandardsIgnoreStart
         $xmp = '<?xpacket begin="' . $this->uniconv->chr(0xfeff) . '" id="W5M0MpCehiHzreSzNTczkc9d"?>' . "\n"
-        . '<x:xmpmeta xmlns:x="adobe:ns:meta/" x:xmptk="Adobe XMP Core 4.2.1-c043 52.372728, 2009/01/18-15:08:04">' . "\n"
+        . '<x:xmpmeta xmlns:x="adobe:ns:meta/"'
+        . ' x:xmptk="Adobe XMP Core 4.2.1-c043 52.372728, 2009/01/18-15:08:04">' . "\n"
         . "\t" . '<rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">' . "\n"
         . "\t\t" . '<rdf:Description rdf:about="" xmlns:dc="http://purl.org/dc/elements/1.1/">' . "\n"
         . "\t\t\t" . '<dc:format>application/pdf</dc:format>' . "\n"
@@ -364,7 +365,9 @@ abstract class MetaInfo extends \Com\Tecnick\Pdf\HTML
         }
 
         // XMP extension schemas
-        $xmp .= "\t\t" . '<rdf:Description rdf:about="" xmlns:pdfaExtension="http://www.aiim.org/pdfa/ns/extension/" xmlns:pdfaSchema="http://www.aiim.org/pdfa/ns/schema#" xmlns:pdfaProperty="http://www.aiim.org/pdfa/ns/property#">' . "\n"
+        $xmp .= "\t\t" . '<rdf:Description rdf:about="" xmlns:pdfaExtension="http://www.aiim.org/pdfa/ns/extension/"'
+            . ' xmlns:pdfaSchema="http://www.aiim.org/pdfa/ns/schema#"'
+            . ' xmlns:pdfaProperty="http://www.aiim.org/pdfa/ns/property#">' . "\n"
         . "\t\t\t" . '<pdfaExtension:schemas>' . "\n"
         . "\t\t\t\t" . '<rdf:Bag>' . "\n"
         . "\t\t\t\t\t" . '<rdf:li rdf:parseType="Resource">' . "\n"
@@ -380,7 +383,8 @@ abstract class MetaInfo extends \Com\Tecnick\Pdf\HTML
         . "\t\t\t\t\t\t\t" . '<rdf:Seq>' . "\n"
         . "\t\t\t\t\t\t\t\t" . '<rdf:li rdf:parseType="Resource">' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:category>internal</pdfaProperty:category>' . "\n"
-        . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:description>UUID based identifier for specific incarnation of a document</pdfaProperty:description>' . "\n"
+        . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:description>UUID based identifier'
+        . ' for specific incarnation of a document</pdfaProperty:description>' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:name>InstanceID</pdfaProperty:name>' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:valueType>URI</pdfaProperty:valueType>' . "\n"
         . "\t\t\t\t\t\t\t\t" . '</rdf:li>' . "\n"
@@ -401,13 +405,15 @@ abstract class MetaInfo extends \Com\Tecnick\Pdf\HTML
         . "\t\t\t\t\t\t\t\t" . '</rdf:li>' . "\n"
         . "\t\t\t\t\t\t\t\t" . '<rdf:li rdf:parseType="Resource">' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:category>internal</pdfaProperty:category>' . "\n"
-        . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:description>Amendment of PDF/A standard</pdfaProperty:description>' . "\n"
+        . "\t\t\t\t\t\t\t\t\t"
+        . '<pdfaProperty:description>Amendment of PDF/A standard</pdfaProperty:description>' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:name>amd</pdfaProperty:name>' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:valueType>Text</pdfaProperty:valueType>' . "\n"
         . "\t\t\t\t\t\t\t\t" . '</rdf:li>' . "\n"
         . "\t\t\t\t\t\t\t\t" . '<rdf:li rdf:parseType="Resource">' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:category>internal</pdfaProperty:category>' . "\n"
-        . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:description>Conformance level of PDF/A standard</pdfaProperty:description>' . "\n"
+        . "\t\t\t\t\t\t\t\t\t"
+        . '<pdfaProperty:description>Conformance level of PDF/A standard</pdfaProperty:description>' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:name>conformance</pdfaProperty:name>' . "\n"
         . "\t\t\t\t\t\t\t\t\t" . '<pdfaProperty:valueType>Text</pdfaProperty:valueType>' . "\n"
         . "\t\t\t\t\t\t\t\t" . '</rdf:li>' . "\n"
