+++ test/MetaInfoTest.php
@@ -18,83 +18,6 @@ namespace Test;
 
 use PHPUnit\Framework\Attributes\DataProvider;
 
-class TestablMetaInfo extends \Com\Tecnick\Pdf\Tcpdf
-{
-    public function exposeGetFormattedDate(int $time): string
-    {
-        return $this->getFormattedDate($time);
-    }
-
-    public function exposeGetXMPFormattedDate(int $time): string
-    {
-        return $this->getXMPFormattedDate($time);
-    }
-
-    public function exposeGetOutDateTimeString(int $time, int $oid): string
-    {
-        return $this->getOutDateTimeString($time, $oid);
-    }
-
-    public function exposeGetOutMetaInfo(): string
-    {
-        return $this->getOutMetaInfo();
-    }
-
-    public function exposeGetEscapedXML(string $str): string
-    {
-        return $this->getEscapedXML($str);
-    }
-
-    public function exposeGetOutXMP(): string
-    {
-        return $this->getOutXMP();
-    }
-
-    public function exposeGetOutViewerPref(): string
-    {
-        return $this->getOutViewerPref();
-    }
-
-    public function exposeGetPageBoxName(string $name): string
-    {
-        return $this->getPageBoxName($name);
-    }
-
-    public function exposeGetPagePrintScaling(): string
-    {
-        return $this->getPagePrintScaling();
-    }
-
-    public function exposeGetDuplexMode(): string
-    {
-        return $this->getDuplexMode();
-    }
-
-    public function exposeGetBooleanMode(string $name): string
-    {
-        return $this->getBooleanMode($name);
-    }
-
-    public function exposeGetProducer(): string
-    {
-        return $this->getProducer();
-    }
-}
-
-class TestableObjPageForMetaInfo extends \Com\Tecnick\Pdf\Page\Page
-{
-    /**
-     * @var array<string, string>
-     */
-    public array $CropBox = [
-        'MediaBox' => 'MediaBox',
-    ];
-
-    public function __construct()
-    {
-    }
-}
-
 class MetaInfoTest extends TestUtil
 {
     protected function getTestObject(): \Com\Tecnick\Pdf\Tcpdf
